+++ src/types/city.ts
@@ -1,0 +1,11 @@
+import { citys } from '../const';
+import { TMapLocation } from './map-location';
+
+type TCityName = typeof citys[number];
+
+type TCity = {
+  location: TMapLocation;
+  name: TCityName;
+};
+
+export type { TCity };
