+++ src/types/host.ts
@@ -1,0 +1,8 @@
+type THost = {
+  avatarUrl: string;
+  id: number;
+  isPro: boolean;
+  name: string;
+};
+
+export type { THost };
