+++ src/types/map-location.ts
@@ -1,0 +1,7 @@
+type TMapLocation = {
+  latitude: number;
+  longitude: number;
+  zoom: number;
+};
+
+export type { TMapLocation };
